Edit-menu/clipboard for opponent history. v2.1.0.0
Uses clipboard. Edit-menu with cut, copy and paste for handing opponents list.  Handy if you have matches in CSV format in text file, so you can now paste them. Paste does not check that matches are in correct format.
Error message if illegal thinking time. That is possible with CSV format. Time in minutes should be a number.